+++ beltradar/api/helpers/core.py
@@ -13,7 +13,7 @@ logger = AppLogger(get_extension_logger(__name__), __title__)
 
 def get_owner_or_none(
     request, character_id
-) -> tuple[bool, models.BeltSurveySession | None]:
+) -> tuple[bool | None, models.BeltSurveySession | None]:
     """Get Character and check permissions"""
     perms = True
 
@@ -22,7 +22,7 @@ def get_owner_or_none(
             owner__profile__main_character__character_id=character_id
         ).first()
         if not survey_session:
-            return False, None
+            return None, None
     except ValueError:
         return None, None
 
@@ -35,15 +35,13 @@ def get_owner_or_none(
 
 def get_public_id_or_none(
     request, public_id
-) -> tuple[bool, models.BeltSurveySession | None]:
+) -> tuple[bool | None, models.BeltSurveySession | None]:
     """Get Survey Session and check permissions"""
     perms = True
 
     try:
         survey_session = models.BeltSurveySession.objects.get(public_id=public_id)
-    except ObjectDoesNotExist:
-        return False, None
-    except ValueError:
+    except (ObjectDoesNotExist, ValueError):
         return None, None
 
     # check access
@@ -55,7 +53,7 @@ def get_public_id_or_none(
 
 def get_belt_timer_or_none(
     request, character_id: int
-) -> tuple[bool, models.BeltTimer | None]:
+) -> tuple[bool | None, models.BeltTimer | None]:
     """Get Belt Timer and check permissions"""
     perms = True